backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
whitenoise==6.12.0
